requirements.txt
Flask==2.3.4
Flask-Login==0.6.3
Flask-Cors==3.1.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
python-dateutil==2.8.2
reportlab==4.1.3
# اختياري: مكتبة خفيفة بديل pandas
panda==0.3.1
